Pull out metrics code into a separate module from components
